chore: Cleanup duplicated code
