+++ src/services/ipmfs/backend.rs
@@ -25,7 +25,6 @@ use serde::Deserialize;
 
 use super::dir_stream::DirStream;
 use super::error::parse_error;
-use super::error::parse_json_deserialize_error;
 use crate::ops::*;
 use crate::raw::*;
 use crate::*;
