+++ src/services/ipmfs/dir_stream.rs
@@ -20,7 +20,6 @@ use serde::Deserialize;
 
 use super::backend::IpmfsBackend;
 use super::error::parse_error;
-use super::error::parse_json_deserialize_error;
 use crate::raw::*;
 use crate::ObjectMetadata;
 use crate::ObjectMode;
