+++ src/services/s3/backend.rs
@@ -46,7 +46,6 @@ use serde::Serialize;
 
 use super::dir_stream::DirStream;
 use super::error::parse_error;
-use super::error::parse_xml_deserialize_error;
 use crate::ops::*;
 use crate::raw::*;
 use crate::*;
